feat: implement immediate SMS sync for premium users
Refactored SMS synchronization logic into a reusable `SmsCloudSynchronizer` class.
Updated `SmsSyncTrigger` to launch immediate synchronization coroutines for premium users, bypassing WorkManager latency while retaining it as a fallback.
Introduced `ApplicationScope` via Hilt for lifecycle-safe coroutine execution.
Split WorkManager tasks into targeted (per-thread) and full sync queues to improve concurrency and responsiveness.
