+++ androidApp/src/main/java/com/pulselink/di/CoroutineScopesModule.kt
@@ -1,0 +1,22 @@
+package com.pulselink.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import javax.inject.Named
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object CoroutineScopesModule {
+
+    @Provides
+    @Singleton
+    @Named("ApplicationScope")
+    fun provideApplicationScope(): CoroutineScope =
+        CoroutineScope(SupervisorJob() + Dispatchers.Default)
+}
